Add option to shrink or expand canvas size
This is handy if for some reason some elements of the chart are cut off.
This typically happens when trying to place the legend outside the plot
area. With this feature one can expand the size of the chart, both in
height and width so that this can be fixed.